rand_distr: add module fisher_f
